// frontend/src/constants/enums.ts
export enum TransactionType {
  INCOME = 'income',
  EXPENSE = 'expense',
  SAVING = 'saving',
}

export enum FrequencyType {
  DAILY = 'zilnic',
  WEEKLY = 'săptămânal',
  MONTHLY = 'lunar',
  YEARLY = 'anual',
}

export enum CategoryType {
  INCOME = 'VENITURI',
  EXPENSE = 'CHELTUIELI',
  SAVING = 'ECONOMII',
}